ESP: make sure the rs_count field does not get included in the netif struct.
The linkoutput function which is linked to the eagle netif is in the
closed source library, so the layout of the netif struct must be equal
to that of what's used in Espressif's version.

The rs_count has to to with IPv6 although IPv6 is not enabled, it's
inclusion in the struct is an error anyway.